Switch dependencies to peerDependencies for Angular & Angular libraries
Using peer dependencies should reduce / eliminate duplicate versions of
modules when this package is included in an application, which has proved
to be a recurring problem especially for Angular libraries.

In order to be able to build the library we still need these modules
available locally, so they are now declared twice, in both peerDependencies
and devDependencies, which should make them available when we build the library
but prevent NPM / Yarn from installing them when we include this library
as a dependency in another application.

Connects to #30