Rework model filtering and update behavior
Instead of updating the model selection every time the modal is closed,
update the selection when the update button is clicked. This prevents
errors related to trying to update the model selection when the parent
chart has been removed from the DOM.

Also, expose a public method to close the modal. This can be used to
close the modal without triggering any updates to the models.

Refs https://github.com/azavea/temperate/issues/660